Add files via upload
This is an Airline Reservation System project for the learner of Vb.net. The system has 2 sides of users which are the Admin and the User. The Admin user is in charge of managing the Aircraft, Destinations, Scheduling, and Reports. The Regular User will be the one who can reserve a flight for the passengers and also the one who can cancel the reservation of the passengers. This Project using SQL Server for the database. The system is user-friendly and has a simple user interface.